Default implementation for addGroupWithConfigurationParams
Use the existing addGroupWithConfig API implementation
which is used to setup a P2P group owner or join a group as a group client
with the specified configuration(SSID, password, band/group).
New modified API called addGroupWithConfigurationParams is added
to include the authentication key management used to setup a connection.

Bug: 297426719
Test: Build successfully
Change-Id: I1d8940e3b94a4a34959c57ef32660dc7f3667ec7